requirements.txt
numpy<2
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
PyMySQL==1.1.1
python-dotenv==1.0.1
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
